refactor lookup components
