+++ awx/ui_next/src/components/Lookup/InstanceGroupsLookup.jsx
