+++ awx/ui_next/src/components/Lookup/InventoriesLookup.jsx
