+++ awx/ui_next/src/components/Lookup/index.js
@@ -1,1 +1,3 @@
 export { default } from './Lookup';
+export { default as InstanceGroupsLookup } from './InstanceGroupsLookup';
+export { default as InventoriesLookup } from './InventoriesLookup';
